+++ app/controllers/corrections_controller.rb
@@ -65,18 +65,10 @@ class CorrectionsController < ApplicationController
   def update
     @comment = Comment.find_by(id: params[:comment])
 
-
-    if correction_params[:essay_line].empty?
-      redirect_to start_correction_path(@correction.hash_id), error: "A linha deve ser preenchida"
-    elsif correction_params[:text_cut].empty?
-      redirect_to start_correction_path(@correction.hash_id), error: "O recorte de texto deve ser especificado"
-    else
       cc = @correction.correction_comments.create({
         correction_id: @correction.id,
         comment_id: @comment.id,
-        essay_line: correction_params[:essay_line],
-        text_cut: correction_params[:text_cut],
-        penalty: correction_params[:penalty]
+        extended_comment: correction_params[:extended_comment]
       })
         
       if cc.errors.any?
@@ -84,7 +76,6 @@ class CorrectionsController < ApplicationController
       else
         redirect_to start_correction_path(@correction.hash_id)
       end
-    end
 
   end
 
@@ -110,7 +101,7 @@ class CorrectionsController < ApplicationController
   private
 
   def correction_params
-    params.require(:correction).permit(:valid_essay, :essay_line, :text_cut, :final_comment, :penalty)
+    params.require(:correction).permit(:valid_essay, :extended_comment, :final_comment)
   end
 
   def set_correction