+++ db/migrate/20200611141411_add_text_cut_to_correction_comments.rb
@@ -1,5 +1,0 @@
-class AddTextCutToCorrectionComments < ActiveRecord::Migration[6.0]
-  def change
-    add_column :correction_comments, :text_cut, :string, null: false, default: ""
-  end
-end