+++ db/migrate/20200621134150_remove_parts_from_correction_comments.rb
@@ -1,0 +1,7 @@
+class RemovePartsFromCorrectionComments < ActiveRecord::Migration[6.0]
+  def change
+    remove_column :correction_comments, :essay_line, :integer
+    remove_column :correction_comments, :penalty, :integer
+    remove_column :correction_comments, :text_cut, :string
+  end
+end